+++ ff_data_server/src/oanda_api/vendor_api_response.rs
@@ -18,7 +18,7 @@ use ff_standard_lib::standardized_types::resolution::Resolution;
 use ff_standard_lib::standardized_types::subscriptions::{DataSubscription, Symbol, SymbolName};
 use ff_standard_lib::StreamName;
 use crate::oanda_api::api_client::{OandaClient, OANDA_IS_CONNECTED};
-use crate::oanda_api::base_data_converters::oanda_quotebar_from_candle;
+use crate::oanda_api::support_and_conversions::oanda_quotebar_from_candle;
 use crate::oanda_api::download::{generate_url};
 use crate::oanda_api::support_and_conversions::oanda_clean_instrument;
 use crate::oanda_api::support_and_conversions::{add_time_to_date, resolution_to_oanda_interval};
